Break code into modular components
Refactor the existing codebase to enhance modularity while ensuring that the full feature set remains intact. This change aims to improve maintainability and readability of the code.
[skip gpt_engineer]
